Rename base EML directory from eml_2_2_0 to eml
