+++ src/metapype/eml/harness.py
@@ -16,12 +16,12 @@ import json
 
 import daiquiri
 
-from metapype.eml_2_2_0.exceptions import MetapypeRuleError
-from metapype.eml_2_2_0 import export
-from metapype.eml_2_2_0 import evaluate
-from metapype.eml_2_2_0 import names
-from metapype.eml_2_2_0 import rule
-from metapype.eml_2_2_0 import validate
+from metapype.eml.exceptions import MetapypeRuleError
+from metapype.eml import export
+from metapype.eml import evaluate
+from metapype.eml import names
+from metapype.eml import rule
+from metapype.eml import validate
 from metapype.model import mp_io
 from metapype.model.node import Node
