+++ src/metapype/eml/rule.py
@@ -20,8 +20,8 @@ import datetime
 import json
 
 from metapype.config import Config
-from metapype.eml_2_2_0.exceptions import MetapypeRuleError
-from metapype.eml_2_2_0 import names
+from metapype.eml.exceptions import MetapypeRuleError
+from metapype.eml import names
 from metapype.model.node import Node
 
 logger = daiquiri.getLogger('validate: ' + __name__)
