+++ src/metapype/eml/rules.json
